+++ packages/flow-core/src/components/f-div/f-div.ts
@@ -1,4 +1,4 @@
-import { html, unsafeCSS } from "lit";
+import { html, PropertyValueMap, unsafeCSS } from "lit";
 import { property, state } from "lit/decorators.js";
 import { FRoot } from "../../mixins/components/f-root/f-root";
 import eleStyle from "./f-div.scss";
@@ -293,7 +293,10 @@ export class FDiv extends FRoot {
 		 */
 		return html`<slot></slot>${this.loading === "loader" ? html`${unsafeSVG(loader)}` : ""}`;
 	}
-	updated() {
+	protected async updated(
+		changedProperties: PropertyValueMap<any> | Map<PropertyKey, unknown>
+	): Promise<void> {
+		super.updated(changedProperties);
 		if (this.variant === "round") {
 			this.style.borderRadius = `${this.offsetHeight / 2}px`;
 		} else if (this.variant === "curved") {